feat: コーヒー評価を 0.5 刻み Double に（KMP 側）
rating を Int(1..5) → Double(0.5..5.0, 0.0=未評価) に変更。

- domain: CoffeeRecord.rating Double、ObserveVisitedCafesUseCase は >0.0 除外
- coffee-editor: draft/onRatingChanged(Double)、validation を 0.5 刻み判定に
- data-local: coffee_record.rating REAL、Mapper の Long/Int 変換撤去、
  0.5 刻み往復テスト追加
- data-firebase android: fromDocument は Number.toDouble()
- sharedUI: rating 表示追随
- data-model.md: rating 仕様を 0.5 刻みに更新

検証: data-local/domain/map テスト緑、assembleSharedLogicXCFramework
（double rating 確認）、androidApp assembleDebug 成功

+++ shared/data-firebase/src/androidMain/kotlin/com/noricoffee/repository/CoffeeFirestoreMapper.kt
@@ -124,7 +124,7 @@ object CoffeeFirestoreMapper {
         val userId = data["userId"] as? String ?: return null
         val visitedOnStr = data["visitedOn"] as? String ?: return null
         val visitedOn = parseLocalDate(visitedOnStr) ?: return null
-        val rating = (data["rating"] as? Number)?.toInt() ?: return null
+        val rating = (data["rating"] as? Number)?.toDouble() ?: return null
         val notes = data["notes"] as? String ?: return null
         val name = data["name"] as? String ?: return null
         val brewMethodName = data["brewMethod"] as? String ?: return null
